requirements.txt
requests
beautifulsoup4
yaspin
python-dateutil
click